requirements.txt
Django==4.1.3
djangorestframework==3.14.0
drf-writable-nested==0.7.0